Reapply "chore!: remove insertOrUpdateSound and uploadCustomSound Meteor methods (#40476)"
This reverts commit 83f1191d3988beb7646b0c5fa5dbe5f8f5fc1bca.
