+++ .changeset/stale-meals-check.md
@@ -1,0 +1,5 @@
+---
+'@rocket.chat/meteor': major
+---
+
+Removes insertOrUpdateSound and uploadCustomSound Meteor methods
