+++ apps/meteor/server/lib/media/custom-sounds/lib/uploadCustomSound.ts
@@ -2,7 +2,7 @@ import { api } from '@rocket.chat/core-services';
 import type { RequiredField } from '@rocket.chat/core-typings';
 import { CustomSounds } from '@rocket.chat/models';
 
-import type { ICustomSoundData } from '../../../../meteor-methods/media/insertOrUpdateSound';
+import type { ICustomSoundData } from './insertOrUpdateSound';
 import { RocketChatFile } from '../../file';
 import { RocketChatFileCustomSoundsInstance } from '../startup/custom-sounds';
