+++ apps/meteor/server/meteor-methods/index.ts
@@ -37,9 +37,7 @@ import './media/deleteCustomSound';
 import './media/deleteEmojiCustom';
 import './media/getS3FileUrl';
 import './media/insertOrUpdateEmoji';
-import './media/insertOrUpdateSound';
 import './media/listCustomSounds';
-import './media/uploadCustomSound';
 import './media/uploadEmojiCustom';
 import './messages/createDirectMessage';
 import './messages/createDiscussion';
